added frontendBackend interface, renamed datastructure, added colors to edges
