+++ backEnd/src/main/java/abstractDataStructure/AbstractTreeNode.java
@@ -1,4 +1,4 @@
-package abstracttree;
+package abstractDataStructure;
 
 
 import java.util.ArrayList;
