+++ backEnd/src/main/java/abstractDataStructure/TreeStructure.java
@@ -1,4 +1,4 @@
-package abstracttree;
+package abstractDataStructure;
 
 
 /**
