+++ backEnd/src/main/java/genome/StrandEdge.java
@@ -4,7 +4,7 @@ package genome;
  * Created by Matthijs on 24-4-2016.
  */
 
-import abstracttree.Edge;
+import abstractDataStructure.Edge;
 
 /***
  * data container class that stores the start and end of an edge, as well as its weight.
