+++ backEnd/src/test/java/abstractDataStructure/AbstractTreeNodeTest.java
@@ -1,4 +1,4 @@
-package abstracttree;
+package abstractDataStructure;
 
 import org.junit.Before;
 import org.junit.Test;
