+++ backEnd/temp/nodes.csv
@@ -1,0 +1,4 @@
+id,sequence,genomes,refGenome,refCoor
+1,AAAAAAAA,AA,AA,371
+2,A,AA,AA,371
+3,C,AA;BB,AA,371
